FIX : ASA.simplify could generate timestamp after last timestamp in epoch; now fixed

+++ nelpy/utils.py
@@ -42,7 +42,7 @@ def frange(start, stop, step):
     """arange with floating point step"""
     # TODO: this function is not very general; we can extend it to work
     # for reverse (stop < start), empty, and default args, etc.
-    num_steps = np.floor((stop-start)/step)
+    num_steps = int(np.floor((stop-start)/step))
     return np.linspace(start, stop, num=num_steps, endpoint=False)
 
 def spatial_information(ratemap):
